+++ platform/ios/test/MGLMapViewScaleBarTests.m
@@ -58,7 +58,7 @@
     // ... but triggering any camera event will update it.
     self.mapView.zoomLevel = 1;
     [self.mapView layoutIfNeeded];
-    
+
     XCTAssertTrue(CGSizeEqualToSize(scaleBar.intrinsicContentSize, CGSizeZero));
     XCTAssertEqual(scaleBar.alpha, 0);
 
@@ -67,4 +67,38 @@
 
     XCTAssertGreaterThan(scaleBar.alpha, 0);
     XCTAssertFalse(CGSizeEqualToSize(scaleBar.intrinsicContentSize, CGSizeZero));
-}@end
+}
+
+- (void)testScaleBarSizeChanged {
+    
+    // The bar maximum bar width can only limit the bar's total width.
+    // Sometimes we should include some space for last label width, the maximum last label width is 30.
+    // We add this hint value to avoid testing failed when need a extra space for last label.
+    CGFloat scaleBarLastLabelWidthHint = 30.0f;
+    
+    self.mapView.scaleBarMaximumWidthRatio = 0.5;
+    UIView *scaleBar = self.mapView.scaleBar;
+    scaleBar.hidden = NO;
+    
+    self.mapView.zoomLevel = 15;
+    [self.mapView layoutIfNeeded];
+    
+    XCTAssertLessThanOrEqual(scaleBar.intrinsicContentSize.width, self.mapView.frame.size.width/2 + scaleBarLastLabelWidthHint);
+    
+    self.mapView.zoomLevel = 10;
+    [self.mapView layoutIfNeeded];
+    XCTAssertLessThanOrEqual(scaleBar.intrinsicContentSize.width, self.mapView.frame.size.width/2 + scaleBarLastLabelWidthHint);
+    
+    CGRect frame = self.mapView.frame;
+    frame.size = CGSizeMake(frame.size.width/2, frame.size.height);
+    self.mapView.frame = frame;
+    [self.mapView layoutIfNeeded];
+    
+    XCTAssertLessThanOrEqual(scaleBar.intrinsicContentSize.width, self.mapView.frame.size.width/2 + scaleBarLastLabelWidthHint);
+    
+    self.mapView.scaleBarMaximumWidthRatio = 0.3;
+    
+    XCTAssertLessThanOrEqual(scaleBar.intrinsicContentSize.width, self.mapView.frame.size.width * 0.3 + scaleBarLastLabelWidthHint);
+}
+
+@end